+++ src/pathPlaning.cpp
@@ -13,7 +13,7 @@ node::node(sc2::Point2D pos,int status, float travelCost,float heuristicToGoal,f
 }
 
 pathPlaning::pathPlaning(CCBot & bot, sc2::Point2D startPos, sc2::Point2D endPos, int mapWidth, int mapHeight, float stepSize):m_bot(bot),m_endPos(endPos),m_stepSize(stepSize),
-	m_map(std::vector<std::vector<std::shared_ptr<node>>>(float(m_bot.Map().width()+1) / stepSize, std::vector<std::shared_ptr<node>>(float(m_bot.Map().height()+1) / stepSize, nullptr)))
+	m_map(std::vector<std::vector<std::shared_ptr<node>>>(static_cast<int>(static_cast<float>(m_bot.Map().width()+1) / stepSize), std::vector<std::shared_ptr<node>>(static_cast<int>(static_cast<float>(m_bot.Map().height()+1) / stepSize), nullptr)))
 {
 	int status = 0;
 	float travelCost = 0.0f;
@@ -28,7 +28,7 @@ const float pathPlaning::calcHeuristic(sc2::Point2D pos) const
 	float xDist = std::abs(pos.x - m_endPos.x);
 	float yDist = std::abs(pos.y - m_endPos.y);
 	//Octile heuristic
-	return std::max(xDist, yDist) + (std::sqrt(2.0) - 1)*std::min(xDist, yDist);
+	return std::max(xDist, yDist) + (std::sqrt(2.0f) - 1.0f)*std::min(xDist, yDist);
 }
 
 const float pathPlaning::calcThreatLvl(sc2::Point2D pos) const
